requirements.txt
fastapi
uvicorn
requests
python-dotenv
networkx
matplotlib
